Class Implementation + Design Tweaks
Tweaked designs so they stay up to date with class implementations

Implemented and Tested the following classes:

1. Transform (Rotation doesn't work, will need looking into later)

+++ SDL Engine/Source/Mathematics/Maths.h
@@ -3,7 +3,7 @@
 struct Maths
 {
 public:
-	static const double PI;
+	static const float PI;
 
 public:
 	static float DegToRad(float deg);
